JCR-2229: Improve behaviour when 'too many open files' occurs
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@799595 13f79535-47bb-0310-9956-ffa450edef68

+++ jackrabbit-core/src/main/java/org/apache/jackrabbit/core/query/lucene/directory/FSDirectoryManager.java
@@ -73,11 +73,15 @@ public class FSDirectoryManager implements DirectoryManager {
                 return pathname.isDirectory();
             }
         });
-        String[] names = new String[dirs.length];
-        for (int i = 0; i < dirs.length; i++) {
-            names[i] = dirs[i].getName();
+        if (dirs != null) {
+            String[] names = new String[dirs.length];
+            for (int i = 0; i < dirs.length; i++) {
+                names[i] = dirs[i].getName();
+            }
+            return names;
+        } else {
+            throw new IOException("listFiles for " + baseDir.getPath() + " returned null");
         }
-        return names;
     }
 
     /**
@@ -90,10 +94,15 @@ public class FSDirectoryManager implements DirectoryManager {
             return true;
         }
         // delete files first
-        for (File file : directory.listFiles()) {
-            if (!file.delete()) {
-                return false;
+        File[] files = directory.listFiles();
+        if (files != null) {
+            for (File file : files) {
+                if (!file.delete()) {
+                    return false;
+                }
             }
+        } else {
+            return false;
         }
         // now delete directory itself
         return directory.delete();
